+++ packages/react/src/misc/index.ts
@@ -2,6 +2,7 @@ export * from './useCurrencies';
 export * from './useInvitedProfiles';
 export * from './useInviteWallets';
 export * from './useModuleMetadata';
+export * from './useResolveAddress';
 export * from './useWasWalletInvited';
 
 export type { InvitedResult } from '@lens-protocol/api-bindings';