+++ packages/react-native/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @lens-protocol/react-native
 
+## 0.1.0-alpha.4
+
+### Minor Changes
+
+- 481e1d7aa: **feat:** adds `useResolveAddress` hook
+
+### Patch Changes
+
+- c074c4ded: **fixed:** issue #777
+- fa944c274: **fixed:** TSDocs typos/mistakes
+- Updated dependencies [481e1d7aa]
+- Updated dependencies [c074c4ded]
+- Updated dependencies [fa944c274]
+  - @lens-protocol/react@2.0.0-alpha.27
+  - @lens-protocol/api-bindings@0.11.0-alpha.24
+
 ## 0.1.0-alpha.3
 
 ### Minor Changes